Moved SectionBanner component into MovieDetailsTile where it's used
