+++ src/common/tiles/MovieDetailsTile/banner/MovieBanner/styled.js
