+++ src/common/tiles/MovieDetailsTile/banner/SectionBanner/index.js
