+++ src/common/tiles/MovieDetailsTile/banner/SectionBanner/styled.js
